Move raster bui to raster path, add real raster test
