+++ cffdrs/raster/__init__.py
@@ -1,0 +1,1 @@
+""" Module that houses raster based FWI calculations """
